update the pr according to the reviews
